Implemented naive unit flocking. Layer unit data structure needs to be implemented using a spatial data structure to improve performance.

+++ C++ Library/PathfindingLib/PathfindingLib/Pathfinding.cpp
@@ -14,8 +14,8 @@ double HeuristicCostEstimate(PGraphNode* a, PGraphNode* b) {
 	return EuclideanDistance(a->position, b->position);
 }
 
-std::vector<blaze::StaticVector<double, 3UL>>* reconstruct_path(std::unordered_map<blaze::StaticVector<int, 3UL>, PGraphNode*, PGraphNode::IndexHash>* came_from, PGraphNode* final_node) {
-	std::vector<blaze::StaticVector<double, 3UL>>* result = new std::vector<blaze::StaticVector<double, 3UL>>;
+std::vector<blaze::StaticVector<double, 3>>* reconstruct_path(std::unordered_map<blaze::StaticVector<int, 3>, PGraphNode*, PGraphNode::IndexHash>* came_from, PGraphNode* final_node) {
+	std::vector<blaze::StaticVector<double, 3>>* result = new std::vector<blaze::StaticVector<double, 3>>;
 	PGraphNode* current = final_node;
 	result->push_back(current->position);
 	while (came_from->find(current->index) != came_from->end()) {
@@ -26,15 +26,15 @@ std::vector<blaze::StaticVector<double, 3UL>>* reconstruct_path(std::unordered_m
 	return result;
 }
 
-std::vector<blaze::StaticVector<double, 3UL>>* Pathfinding::a_star(PGraph * graph, blaze::StaticVector<double, 3UL> from, blaze::StaticVector<double, 3UL> to)
+std::vector<blaze::StaticVector<double, 3>>* Pathfinding::a_star(PGraph * graph, blaze::StaticVector<double, 3> from, blaze::StaticVector<double, 3> to)
 {
 	PGraphNode* from_node = graph->NodeAt(from);
 	PGraphNode* to_node = graph->NodeAt(to);
 
-	std::unordered_map<blaze::StaticVector<int, 3UL>, float, PGraphNode::IndexHash> f_score;
-	std::unordered_set<blaze::StaticVector<int, 3UL>, PGraphNode::IndexHash> closed_set;
-	std::unordered_map<blaze::StaticVector<int, 3UL>, PGraphNode*, PGraphNode::IndexHash> came_from;
-	std::unordered_map<blaze::StaticVector<int, 3UL>, float, PGraphNode::IndexHash> g_score;
+	std::unordered_map<blaze::StaticVector<int, 3>, float, PGraphNode::IndexHash> f_score;
+	std::unordered_set<blaze::StaticVector<int, 3>, PGraphNode::IndexHash> closed_set;
+	std::unordered_map<blaze::StaticVector<int, 3>, PGraphNode*, PGraphNode::IndexHash> came_from;
+	std::unordered_map<blaze::StaticVector<int, 3>, float, PGraphNode::IndexHash> g_score;
 
 	NodeMinHeap open_set(&f_score);
 	f_score.emplace(from_node->index, HeuristicCostEstimate(from_node, to_node)); // must insert into f_score before open_set
@@ -55,7 +55,7 @@ std::vector<blaze::StaticVector<double, 3UL>>* Pathfinding::a_star(PGraph * grap
 				continue;
 			}
 
-			float tentative_g_score = g_score.at(current->index) + EuclideanDistance(current->position, neighbor->position);
+			float tentative_g_score = g_score.at(current->index) + static_cast<float>(EuclideanDistance(current->position, neighbor->position));
 			float current_g_score = FLT_MAX;
 			if (g_score.find(neighbor->index) != g_score.end()) {
 				current_g_score = g_score.at(neighbor->index);
@@ -66,7 +66,7 @@ std::vector<blaze::StaticVector<double, 3UL>>* Pathfinding::a_star(PGraph * grap
 
 			came_from[neighbor->index] = current;
 			g_score[neighbor->index] = tentative_g_score;
-			f_score[neighbor->index] = g_score.at(neighbor->index) + HeuristicCostEstimate(neighbor, to_node);
+			f_score[neighbor->index] = g_score.at(neighbor->index) + static_cast<float>(HeuristicCostEstimate(neighbor, to_node));
 
 			if (open_set.Contains(neighbor)) {
 				open_set.Update(neighbor);
@@ -77,5 +77,5 @@ std::vector<blaze::StaticVector<double, 3UL>>* Pathfinding::a_star(PGraph * grap
 		}
 	}
 
-	return new std::vector<blaze::StaticVector<double, 3UL>>;
+	return new std::vector<blaze::StaticVector<double, 3>>;
 }
